Move LarkyUnittest out of test scope
Move LarkyUnittest module to src/nativelibs instead of keeping it in test. This
allows us to keep a consistent interface of all nativelibs in a folder.

As a result, we have to move junit dependency out of the test scope and into
compile.

Also, I suppressed some superflous inspections for LarkyUnittest because we
aren't actually using the inner classes for dependency injection or different runners.