Implement JSR223 Invocable interface
Allows Larky to execute functions via the JSR223 Invocable interface. This
allows us to evaluate a script and then execute a function that is exposed
via the script. This should be helpful in executing the `process(message, ctx)`
interface that is exposed via user-defined VGS operations.

This dramatically simplified the LarkyCompiledScript code.

After discovering the `AbstractScriptEngine` class, the `LarkyScriptEngine`
class was simplified to extend `AbstractScriptEngine` to avoid boiler plate
code.

Also, as part of this change, the following has been implemented:

- changed the ParsedStarFile to implement the StarFile interface.
- Introduced the `sum()` function into the python builtins