Update Starlark to Bazel Master #543f6e6 (#24)
* Update Starlark to Bazel Master #543f6e6

- starlark: delete Module.getTransitiveBindings and get Make
clients spell out what they mean instead of just looking for key/value pairs in
random places. Starlark is a statically scoped language, after all.
- starlark: implement "flat frames" optimization  …  The resolver now statically maps each local variable to a small integer, and the set of local bindings (including from comprehensions) is accumulated into each Resolver.Function. The evaluator uses an array instead of a LinkedHashMap to represent the environment, and no longer needs a hack to save/restore the values of comprehensions. Also, added a test of the resolver.
- starlark: always record resolver information in syntax tree  …  FileOptions.recordScope(false) is no longer needed.
- Starlark: more 64-bit integer operations without round-trip to BigInteger
Implement faster integer operations for integers within signed 64-bit range:
* comparison
* addition
* subtraction Note JDK has `Math.addExact` function; it cannot cannot be used
directly, because thrown exception on overflow results in high performance
slowdown (about 3x on my laptop) because stack trace is collected on overflow.
64-bit integer benchmark is significantly faster.
- starlark: fix ArrayStoreException crash in list.append  …  ...by establishing the invariant that StarlarkList.elems.getClass() == Object[].class. + Test
 starlark: add tests of int / int  …  Also, remove StarklarkInt.divide, which is unused and untested. (int / int is implemented by first converting operands to float.)
- Starlark: microoptimize str[index]  …  Preallocate single-char ASCII strings. ``` def test():  for i in range(1000000):  s = "abcdefghijklmnopqrstuvwxyz"  for j in range(len(s)):  s[j] test() ``` ``` A: N=48, r=3.896+-0.475 B: N=48, r=3.516+-0.571 B/A: 0.902 ``` Closes #12490.
- Starlark: optimize 64-bit integer negation  …  Perform negation of 64-bit integer without roundtrip to `BigInteger`. Not a big deal, but it may be a performance issue in certain cases.
- Starlark: StarlarkCallable.allowReturnNones as Nullable. Functional change is in StarlarkMethodProcessor. Other changes are adding missing Nullable annotations (or removing unnecessary allowReturnNones, for example when returning None or primitive types). At this point @nullable annotation does almost nothing, but if we decide to use nullness checker, the checker would guarantee that a method may return null iff annotated.
- Starlark: only nested classes of StarlarkInt are allowed to inherit it  …  Minor code cleanup

* Forgot to checkin the patched LarkyScript

* Fix pulling from globals and predeclared

* Update Starlark to [latest](https://github.com/bazelbuild/bazel/commit/8974ba2518603710c9e602d6d8518b7494eadc09)

starlark resolver: implement "flat globals" optimization

This change reworks the Resolver.Module interface (which abstracts
eval.Module) so that each global is statically mapped to a small integer,
so that they can be represented during execution as a flat array,
not a hash table, similar to the treatment of locals.

It also resolves PREDECLARED and UNIVERSE references distinctly, so that
we needn't try both maps during execution. And it sets the stage for
those maps to be treated as flat arrays too.

Furthermore, it materializes for the first time a "file-local block",
which will be used to hold bindings created by load statements, when
we make loads bind locally and eliminate the getExportedGlobals hack.

The Resolver.Module interface no longer requires the implementation to
enumerate all the defined names; names are looked up on demand as
references are encountered in the program during resolution, and
Bindings are created for them. This reduces unnecessary allocation and
copying when a small program is executed in a big environment.
The resolver still provides spelling suggestions over all available
names; it does this by having the Module implementation provide
the set of top-level candidates lazily: only when resolve() fails.

* Sync with upstream Bazel by restoring files

* Fix failing JSR223 ScriptEngineFactory::getNames() test