Updated semantics (__name__ instead of __class__)
Using __name__ on the class and __class__ as the type instantiator is the *right* pythonic semantics. Instead of __class__ being a string, it should be the class factory. __name__ is the actual name of the "structure" and this change updates the semantics for this.

This has a fallback behavior to avoid having to make this a breaking change as well.