Overhaul Larky struct module
Updated the Struct Module and passed all the relevant tests from
https://github.com/python/cpython/blob/b08c48e/Lib/test/test_struct.py

The struct library, https://docs.python.org/3/library/struct.html, is a very
pervasive library in the Python eco-system and has lots of usage in the
cryptography libraries. The previous implementation was very naive and was
barely enough to pass simple tests but as more sophisticated use cases become
wide-spread, having it properly tested and implemented allows for bug-free
guarantees.