Remove Quarkus
We put Quarkus in Larky to create a graalvm compatible native image
but we did not use any kind of webserver support. We just wanted to
use the command-line client. It turns out this is something we can
easily achieve with a much more dependency friendly utility called
picocli and it has built-in support for command-line graalvm native
applications!

This PR does just that and keeps the interface exactly the same.

commit-id:30d07b40