Step 3 of introducing types in Larky:
In Starlark, `Starlark.type(object)` return the *string* representing the type class of the passed `object`. We introduce LarkySemantics which allows us to introduce Python-like semantics in Larky, such that if they are turned off, Larky will behave like Starlark and if these flags are flipped, Larky will behave like Python.

This is important because in Python, methods are invoked from the instance object's *type*, which is retrieved by: `instance.__class__` *OR* `type(cls)`. This diff allows us to pass in a `LarkySemantics` in our LarkyEvaluator and LarkyScript.

In addition (and unrelated to this diff), use Lombok `@Getter`s to avoid boiler plate.

commit-id:1c36312c