Step 6 of introducing types in Larky:
Introduce LarkyType.

In Python, the base type is `object` and in Python, a type is an object. We introduce this interface to mimic setting up the described hierarchy.

For LarkyType, we note two methods named `lookup()` and `lookupForSuper()`.

The default behavior for attribute access is to get, set, or delete the attribute from an object’s dictionary.
For instance, `a.x` has a lookup chain starting with `a.__dict__['x']``, then `type(a).__dict__['x']``, and
continuing through the base classes of type(a) excluding metaclasses.

However, for Class Bindings, which is the equivalent of `LarkyType`:

 - If binding to a class, A.x is transformed into the call: `A.__dict__['x'].__get__(None, A)`.

Special attributes for a Class:

__name__ - The class name.
__dict__ - The dictionary containing the class’s namespace.
__bases__ - A tuple containing the base classes, in the order of their occurrence in the base class list.
__mro__ - This attribute is a tuple of classes that are considered when looking for base classes during method resolution.

More here:
- https://docs.python.org/3/library/stdtypes.html?highlight=__mro__#special-attributes
- https://docs.python.org/3/c-api/typeobj.html?highlight=__base__#tp-slots

commit-id:9896c9f8