Step 11 of introducing types in Larky:
Introduce descriptor types! 😲🔥🍾 In Python, the developer *controls the dot* and this makes up the entire attribute access chain that makes up the Python data model. Descriptors are a powerful, general purpose protocol. They are the mechanism behind properties, methods, static methods, class methods, and super(). They are used throughout Python itself.

In general, a descriptor is an attribute value that has one of the methods in the descriptor protocol. Those methods are `__get__()`, `__set__()`, and `__delete__()`. If any of those methods are defined for an attribute, it is said to be a descriptor.

The default behavior for attribute access is to `get`, `set`, or `delete` the attribute from an object’s dictionary. For instance, `a.x` has a lookup chain starting with `a.__dict__['x']`, then `type(a).__dict__['x']`, and continuing through the method resolution order of `type(a)`. If the looked-up value is an object defining one of the descriptor methods, then Python may override the default behavior and invoke the descriptor method instead. Where this occurs in the precedence chain depends on which descriptor methods were defined.

Descriptor protocol:

- `descr.__get__(self, obj, type=None) -> value`
- `descr.__set__(self, obj, value) -> None`
- `descr.__delete__(self, obj) -> None`

That's it! Define any of these methods and an object is considered a descriptor and can override default behavior upon being looked up as an attribute.

If an object defines `__set__()` **OR** _`_delete__()`, it is considered a `data descriptor` -- see `LarkyDataDescriptor`. Descriptors that only define `__get__()` are called non-data descriptors -- see `LarkyNonDataDescriptor`.

commit-id:98d567ad