Step 13 of introducing types in Larky:
We introduce LarkyFunction and LarkyMethod. These are important in Python and are now first class in Larky.

The previous PRs were building to OO features in Larky and now we have reached this crescendo 🎵🙌.

Python’s object-oriented features are built upon a function based environment. Using non-data descriptors, the two are merged seamlessly.

- Functions stored in class dictionaries get turned into methods when invoked. (LarkyFunction)
- Methods only differ from regular functions in that the object instance is prepended to the other arguments. (LarkyMethod)

By convention, the instance is called self but could be called this or any other variable name.

Classes! Now, when we set a function via `types.new_class`, we automatically check to see if there's a StarlarkFunction (just a normal function) and we then wrap it in a LarkyFunction which understands how to bind itself to an instance.

This PR is accompanied by the implementation of the `__getattribute__` Python method, effectively the starting point of the descriptor protocol.

The expression `obj.x` looks up the attribute `x` in the chain of namespaces for `obj`. If the search finds a descriptor outside the instance `__dict__`, its `__get__()` method is invoked according to the precedence rules listed below.

Instance lookup scans through a chain of namespaces:
 1. giving data descriptors the highest priority
 2. followed by instance variables
 3. then, non-data descriptors
 4. then class variables,
 5. and, lastly `__getattr__()`, if it is provided.

If a descriptor is found for `a.x`, then it is invoked with: `desc.__get__(a, type(a))`.

See more here: https://docs.python.org/3.10/reference/datamodel.html#object.__getattribute__

commit-id:55b700d2