fix: Mimic the semantics in PyCryptodome and translate them to BouncyCastle
In BouncyCastle, if the expected plaintext length is -1
(as opposed to the default in PyCryptodome of 0)
then it means we do not have an expected plain text length.

We add a patch before we invoke `JCrypto` implementation of PKCS1_v1_5#decode.
Since, BouncyCastle assumes that the expected plaintext is actually 0, as
opposed to what PyCryptodome intends, we mimic the behavior.

commit-id:5eaf9705