fix: textwrap without negative and positive lookahead assertions
Python's textwrap uses negative and positive lookahead regular expression assertions which are incompatible with Larky's
regular expression package -- which leverages Google's re2 engine for deterministic run-time behavior.

This diff re-writes and emulates the use of negative and positive lookahead assertions to mimic the functionality needed for Textwrap.

This is needed in the PGP module for armoring and serializing PGP keys.

commit-id:54e3cff6