Step 13 of introducing types in Larky: (#271)
* Step 13 of introducing types in Larky:

We introduce LarkyFunction and LarkyMethod. These are important in Python and are now first class in Larky.

The previous PRs were building to OO features in Larky and now we have reached this crescendo 🎵🙌.

Python’s object-oriented features are built upon a function based environment. Using non-data descriptors, the two are merged seamlessly.

- Functions stored in class dictionaries get turned into methods when invoked. (LarkyFunction)
- Methods only differ from regular functions in that the object instance is prepended to the other arguments. (LarkyMethod)

By convention, the instance is called self but could be called this or any other variable name.

Classes! Now, when we set a function via `types.new_class`, we automatically check to see if there's a StarlarkFunction (just a normal function) and we then wrap it in a LarkyFunction which understands how to bind itself to an instance.

This PR is accompanied by the implementation of the `__getattribute__` Python method, effectively the starting point of the descriptor protocol.

The expression `obj.x` looks up the attribute `x` in the chain of namespaces for `obj`. If the search finds a descriptor outside the instance `__dict__`, its `__get__()` method is invoked according to the precedence rules listed below.

Instance lookup scans through a chain of namespaces:
 1. giving data descriptors the highest priority
 2. followed by instance variables
 3. then, non-data descriptors
 4. then class variables,
 5. and, lastly `__getattr__()`, if it is provided.

If a descriptor is found for `a.x`, then it is invoked with: `desc.__get__(a, type(a))`.

See more here: https://docs.python.org/3.10/reference/datamodel.html#object.__getattribute__

commit-id:55b700d2

* Step 14 of introducing types in Larky: (#272)

* Step 14 of introducing types in Larky:

We introduce SetAttribute and DeleteAttribute to support `object.__setattr__` and `object.__delattr__`. In Larky, this translates to:

```python
- `a.x = 1`
- `setattr(a, x, 1)`
- `delattr(a, x)` # to be implemented
 ```

 This allows us to implement data descriptors (set and delete)!

commit-id:db5fd67c

* Step 15 of introducing types in Larky: (#273)

* Step 15 of introducing types in Larky:

Introduce `super()`!

commit-id:8f028634

* Step 16 of introducing types in Larky: (#274)

* Step 16 of introducing types in Larky:

Introduce `classmethod()` and `staticmethod()` types! And with that, all `test_types.star` pass!

commit-id:be9c46c5

* Step 17 of introducing types in Larky: (#276)

* Step 17 of introducing types in Larky:

- Rename LarkyPyObjectBuiltin => LarkyBaseObjectType, there's no reason it should extend from LarkyPyObject as it is a root base type (not an object).
- Introduce SpecialMethods to allow us to determine quickly if a LarkyProvidedTypeClass overloads a particular operator or introduces a special (`__dunder__` method).
- Introduce `LarkyCollection`, which allows us to encapsulate all the different types of logic needed for operator overloading (similar to SimpleStruct), but this time, it conforms more to Python
commit-id:2510e898

* Step 18 of introducing types in Larky: (#277)

* Step 18 of introducing types in Larky:

fix: Update `unittest.expectedFailure` with exception types

This test ultimately enumerates all application expected Exception classes. Introducing
 `UncheckedEvalException` and `StarlarkEvalWrapper.Exc.RuntimeEvalException` required us
 to update this multi-catch statement as well.

Unrelated to the diff changes is to also remove redundant suppression statements.
commit-id:feed1e3d

* Step 19 of introducing types in Larky: (#278)

* Step 19 of introducing types in Larky:

feature: `LarkyPyObject` now supports overridden `__eq__`

As a future enhancement, we must support the commutative property for the converse. That means if `__ne__` is defined but `__eq__` is not,
we can support `==` (i.e. `__eq__`) as `not __ne__()`.
commit-id:61ffc6cf

* Step 20 of introducing types in Larky: (#279)

* Step 20 of introducing types in Larky:

Introduce a `setItemUnsafe()` method to set an attribute directly on the PyObject internal dictionary.

In python, type `__dict__` is read-only. For objects, it is mutable. This diff implements this behavior.

commit-id:3d67c451

* fix: larky.is_subclass had a typo in variable names that was ultimately a tautology. (#280)

* fix: larky.is_subclass had a typo in variable names that was ultimately a tautology.
commit-id:cf4f8ca3

* chore: Convert DeterministicGenerator to our new type system! (#281)

* chore: Convert DeterministicGenerator to our new type system!
commit-id:dd90ddbf

* fix: Mimic the semantics in PyCryptodome and translate them to BouncyCastle (#282)

* fix: Mimic the semantics in PyCryptodome and translate them to BouncyCastle

In BouncyCastle, if the expected plaintext length is -1
(as opposed to the default in PyCryptodome of 0)
then it means we do not have an expected plain text length.

We add a patch before we invoke `JCrypto` implementation of PKCS1_v1_5#decode.
Since, BouncyCastle assumes that the expected plaintext is actually 0, as
opposed to what PyCryptodome intends, we mimic the behavior.

commit-id:5eaf9705

* fix: textwrap without negative and positive lookahead assertions (#283)

* fix: textwrap without negative and positive lookahead assertions

Python's textwrap uses negative and positive lookahead regular expression assertions which are incompatible with Larky's
regular expression package -- which leverages Google's re2 engine for deterministic run-time behavior.

This diff re-writes and emulates the use of negative and positive lookahead assertions to mimic the functionality needed for Textwrap.

This is needed in the PGP module for armoring and serializing PGP keys.

commit-id:54e3cff6

* fix: for all our hashes, fix a bug that was checking for falsey values which would fail when the input was an empty byte type (byte string or bytearray). (#284)

* fix: for all our hashes, fix a bug that was checking for falsey values which would fail when the input was an empty byte type (byte string or bytearray).
commit-id:8eb9bfe8

* chore: Add `__class__` and `__name__` to the RSA mutablestruct so that we can comply with our `larky.is_instance()` protocols. (#285)

* chore: Add `__class__` and `__name__` to the RSA mutablestruct so that we can comply with our `larky.is_instance()` protocols.
commit-id:392e70b9

* Introduce PGP support with a passing test! ✅ (#286)

* Introduce PGP support with a passing test! ✅

commit-id:30bef1df

* All tests now pass! We now support PGP in Larky! 🍾🎆 (#287)

* All tests now pass! We now support PGP in Larky! 🍾🎆

1. Cleanup by removing __init__2.star
2. Switch from '%02X' format string to using `hexlify`
commit-id:fe2b02c1

* Remove docker test for now

Co-authored-by: Andrew <a.slepakurov@gmail.com>

Co-authored-by: Andrew <a.slepakurov@gmail.com>

Co-authored-by: Andrew <a.slepakurov@gmail.com>

Co-authored-by: Andrew <a.slepakurov@gmail.com>

Co-authored-by: Andrew <a.slepakurov@gmail.com>

Co-authored-by: Andrew <a.slepakurov@gmail.com>

Co-authored-by: Andrew <a.slepakurov@gmail.com>

Co-authored-by: Andrew <a.slepakurov@gmail.com>

Co-authored-by: Andrew <a.slepakurov@gmail.com>

Co-authored-by: Andrew <a.slepakurov@gmail.com>

Co-authored-by: Andrew <a.slepakurov@gmail.com>

Co-authored-by: Andrew <a.slepakurov@gmail.com>

Co-authored-by: Andrew <a.slepakurov@gmail.com>

Co-authored-by: Andrew <a.slepakurov@gmail.com>

Co-authored-by: Andrew <a.slepakurov@gmail.com>

Co-authored-by: Andrew <a.slepakurov@gmail.com>