fix: Make all the tests pass
The key issue was just sloppy porting to Larky. Comparisons can be done via the `operator` module.

In this commit, is a fix for a *HUGE* issue in `LarkyDigest` that was checking to see if `LarkyDigest` was of instance `Memoable` *INSTEAD* of the underlying `ExtendedDigest` from bouncycastle. This caused a failure where repeated invocations of `hexdigest()` or `digest()` would produce the wrong digest value. The first attempt to fix this was in https://github.com/verygoodsecurity/starlarky/commit/020fde286b4fca6e16cbbb44b81fd3584833f07a for reported issue #179.

commit-id:8858cdb3