🌈 Python-JOSE's jwe is now ported to Larky with passing tests from #296 💪 (#297)
* chore: whitespace cleanup, better error messages and more conversions from `safe()` to using the `Result.Ok().map()` pattern.

commit-id:8520c44f

* chore: update comments on LarkyDescriptor and LarkyFunction to describe a possible optimization.

commit-id:1ec7c828

* feature: introduce failing jws test from @mottersheadt

commit-id:8c2a247a

* feature: port python-jose's jws.py to jws.star for larky

commit-id:eb8d99a7

* feature: fix first test in `test_jws.star`

It was failing because in PR #295 or PR #250, we did not update the SUPPORTED algorithms.

commit-id:c75f4092

* feature: introduce a non-pycryptodome compatible ECC curve, a Koblitz curve, SecP256K1Curve to fix the ECC test.

The ECC key was generated by @mottersheadt from https://cloud.google.com/iot/docs/how-tos/credentials/keys via:

- `openssl ecparam -genkey -name prime256v1 -noout -out ec_private.pem`
- `openssl ec -in ec_private.pem -pubout -out ec_public.pem`

The "k" in sepc256k1 stands for Koblitz and the "r" in sepc256r1
stands for random. A Koblitz elliptic curve has some special properties
that make it possible to implement the group operation more efficiently.

secp256k1 was almost never used before Bitcoin became popular, but it is
now gaining in popularity due to its several nice properties. Most
commonly-used curves have a random structure, but secp256k1 was constructed
in a special non-random way which allows for especially efficient
computation. As a result, it is often more than 30% faster than
other curves if the implementation is sufficiently optimized.

Also, unlike the popular NIST curves, secp256k1's constants were selected
in a predictable way, which significantly reduces the possibility that
the curve's creator inserted any sort of backdoor into the curve.

commit-id:0f0c6454

* fix: in our ASN1 implementation, all `toStarlark()` should return `getEncoded()`

This is a continuation of the fixes introduced in #295, but this fixes the *last failing test* in `test_jws.star`. 🍾🎆

commit-id:c43aef7e

* Add example passing tests for verifying signatures without jws

commit-id:9d896504