fix(bug): When using a BufferedBlockCipher, check if there's any further buffer left over before calling doFinal.
In BouncyCastle, we should be using `doFinal` without worrying about
the buffer size or if there's any partial buffer left over to process.
However, when the logic is split between pycryptodome (specifically
for the padding logic) and BouncyCastle, we have to add this check in
to make sure we're not issuing a `reset()` on the internal state of
the block cipher.

In addition, if we were to re-use the cipher (as pycryptodome does if
the caller passes in `msg_len` or `assoc_len`), then the `doFinal`
state can cause incorrect output.

Without this diff that adds a re-entrant init() check as well as a
check to see if there's any more data to process, we will not be able
to verify the MAC on any AEAD cipher that uses a legacy mode cipher
since we will not invoke `doFinal()`. BouncyCastle will (correctly)
attempt to calculate the MAC, however, this part is done in
pycryptodome and as a result, we need to add the check to see if there
is more data left to process.