fix(bug): port over Python's `_namespaces()` function instead of using the overloaded one that mimicked the lxml library
there is one minor difference between this function and the one from python: the larky function properly respects namespaces and will automatically append all referenced namespaces correctly as opposed to only emitting used namespaces. This should generally be OK as it's still user defined. If this continues to become an issue, we can revert the behavior as part of `ElementTree::tostring()` by adding a `python_compat=True` flag.